pyproject.toml
[tool.poetry]
name = "poracle-middleman"
version = "0.1.0"
description = ""
authors = ["Malte <42342921+ccev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
middleman = "main:main"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.4"
"discord.py" = "^2.1.1"
rtoml = "^0.9.0"
pydantic = "^1.10.5"
shapely = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
